+++ backend/cwa_lib/routers/manage_contexts.py
@@ -2,15 +2,21 @@ from fastapi import APIRouter, Depends
 from common.sql_db_async import AsyncSession, async_get_session
 from common.sql_models.api_users import User
 from cwa_lib.app import current_active_user
-from cwa_lib.sql_tables.group_contexts import ManageContextsQuery, ManageContextsQueryResult
+from cwa_lib.pydantic_schemas.generic_table import TableQuery
+from cwa_lib.pydantic_schemas.manage_contexts import ManageContextsQueryResult
+from cwa_lib.sql_tables.group_contexts import ManageContextsTable
 
 router__manage_contexts = APIRouter()
 
 
 @router__manage_contexts.post("/manage_contexts/query", tags=["Manage Contexts"], response_model=ManageContextsQueryResult)
 async def manage_contexts__query(
+    payload: TableQuery,
     user: User = Depends(current_active_user),
     session: AsyncSession = Depends(async_get_session),
     ):
-    result = await ManageContextsQuery(session).query_all_by_group_id(group_id=user.group_id)
+    result = await ManageContextsTable(session).query_all_by_group_id(
+        group_id=user.group_id, 
+        payload=payload
+    )
     return result