+++ frontend/src/tables/CellRenders/TextCell.tsx
@@ -1,0 +1,11 @@
+interface Props {
+  value: boolean;
+}
+
+export default function TextCell({ value }: Props) {
+  return (
+    <textarea className="form-control" rows={5}>
+      {value}
+    </textarea>
+  );
+}